+++ .idea/gradle.xml
@@ -9,8 +9,8 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/ROS2Connection" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/ros2_publisher" />
           </set>
         </option>
         <option name="resolveExternalAnnotations" value="false" />
