+++ ROS2Connection/src/main/cpp/publisher.h
@@ -1,0 +1,14 @@
+//
+// Created by awaty on 2024/09/16.
+//
+
+#ifndef MILKYPUBLISHER_PUBLISHER_H
+#define MILKYPUBLISHER_PUBLISHER_H
+
+
+class publisher {
+
+};
+
+
+#endif //MILKYPUBLISHER_PUBLISHER_H
