--- ros2_publisher/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-    <string name="app_name">ros2_publisher</string>
-</resources>
